Move emscripten_pthread initialization to its own file, out from locale_map.c
